장소 상세보기에서 후기(댓글)  삭제 기능 완료 , (댓글달려있을시 내용만 바뀜, 부모댓글이 모두 삭제되어 내용만 바뀌어있는 상태에서 제일 아래 자식 댓글을 지우면 삭제되있는 위의 부모댓글을까지 같이 삭제)

+++ YOGI/src/main/java/yogi/lendplace/LendplaceController.java
@@ -87,11 +87,23 @@ public class LendplaceController {
     //후기 삭제
     @RequestMapping(value="/lendplace/deleteReview")
     public ModelAndView deleteReview(CommandMap commandMap) throws Exception{
-    	lendplaceService.deleteReview(commandMap.getMap());
-    	return new ModelAndView("redirect:/lendplace/detail?L_NO=" + commandMap.getMap().get("L_NO"));
+    	ModelAndView mv = new ModelAndView("redirect:/lendplace/detail?L_NO=" + commandMap.getMap().get("L_NO"));
+    	Map<String,Object> map = lendplaceService.selectReviewChild(commandMap.getMap());
+    	Integer cnt = Integer.parseInt(map.get("CNT").toString());
+    	if ( cnt == 0) {
+    		Map<String,Object> del = lendplaceService.selectDeletedParent(commandMap.getMap());
+    		Map<String,Object> par = lendplaceService.selectParent(commandMap.getMap());
+    		Integer DEL = Integer.parseInt(del.get("DEL").toString());
+    		Integer PAR = Integer.parseInt(par.get("PAR").toString());
+    		
+    		lendplaceService.deleteReview(commandMap.getMap());
+    		if ( DEL == (PAR-1)) {
+				lendplaceService.deleteGroupReview(commandMap.getMap());
+			}
+        	return mv;	
+        }else {
+        	lendplaceService.updateDeleteFlag(commandMap.getMap());
+    		return mv;
+        }
     }
-
-   
-
-
 }
